+++ SharePointFramework/PortfolioWebParts/src/loc/nb-no.js
@@ -4,7 +4,6 @@ define([], function () {
     FilterBy: 'Filtrer etter',
     DataSourceCategoryLabel: 'Kategori for datakilder',
     DataSourceLabel: 'Datakilde',
-    AddColumnLabel: 'Legg til kolonne',
     SortDescLabel: 'A til Å',
     SortAscLabel: 'Å til A',
     ColumSettingsLabel: 'Kolonneinnstillinger',