+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/MatrixCell/IMatrixCellProps.tsx
@@ -1,0 +1,2 @@
+export interface IMatrixCellProps extends React.HTMLProps<HTMLElement> {
+}