+++ CHANGELOG.md
@@ -1,15 +1,34 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.1.7 - 2020-01-09
+
+### Added
+- RiskMatrix added as separate web part #97
+- RiskMatrix added to project status #172
+- Improved error messages when provisioning new projects #170
+- Including previous budget numbers in new project status #167
+- Better support for the Portfolio administrator role #133
+- Made it possible to work with draft and published versions of project status reports #119
+- Support for copying more than 100 items in CopyListData (up to 500)
+- Support for installing to /teams/ #177
+
+### Fixed
+- Fixed lookups in list 'Prosjektkolonnekonfigurasjon' #142
+- Fixed colors and columns not matching content #134
+- Fixed an issue with invalid web part properties on Oppgaver.aspx #164
+- Added support for currency fields in Portfolio Insights #155
+- Fields with _ in field name doesn't sync to portfolio
+- Persists selection for ListContentSection/ExtensionsSection #182
+
 ## 1.1.6 - 2019-11-14
 
-## Fixed
+### Fixed
 - Fixed a bug with current phase not being displayed in phase web part on project frontpage #149
-
 
 ## 1.1.5 - 2019-11-13
 
-## Added
+### Added
 - Support for PSCredential in Install script #145
 - Added missing resource for choice option (Choice_GtResourceAbsence_Linetasks) #148